Renamed Operation -> ReplyOperation
