+++ tools/util/ReplayOperationManager.hpp
